+++ AssessX/src/main/java/module-info.java
@@ -2,8 +2,6 @@ module com.assessx.assessx {
     requires javafx.controls;
     requires javafx.fxml;
     requires javafx.web;
-    requires org.controlsfx.controls;
-    requires net.synedra.validatorfx;
     requires org.kordamp.ikonli.javafx;
     requires com.google.gson;
     requires java.net.http;
@@ -14,11 +12,9 @@ module com.assessx.assessx {
     opens com.assessx.assessx.controller.pages to javafx.fxml;
     opens com.assessx.assessx.dto to javafx.fxml, com.google.gson;
     opens com.assessx.assessx.session to javafx.fxml;
-    opens com.assessx.assessx.ui to javafx.fxml;
     opens com.assessx.assessx.controller.dialogs to javafx.fxml;
 
     exports com.assessx.assessx;
     exports com.assessx.assessx.api;
     exports com.assessx.assessx.session;
-    exports com.assessx.assessx.ui;
 }